+++ src/store/index.ts
@@ -2,6 +2,7 @@ export * from "./backup";
 export * from "./database";
 export * from "./errors";
 export * from "./import-export";
+export * from "./metadata";
 export * from "./queries";
 export * from "./schema";
 export * from "./syx-codec";